Change how callbacks work
The emulator had a mutable reference to the callbacks structure,
which made it extremely difficult to use, due to the inability to
hold any other references on it and the inability to make this
part of a structure.

To get rid of the reference, the callback is now passed as a
parameter to the try_io_emulation method, replacing the PVOID
context (helping to avoid unsafe casts).